requirements.txt
pillow==11.0.0
requests==2.32.3
numpy==1.26.4
